+++ api/src/handlers/utility.rs
@@ -1,0 +1,1 @@
+// write utility method
